Various Files
fireproj_getdata_for_ml is just getting the past years worth of weather and fire data in a format to train the ML model. The code is ugly and not well documented because there was a lot of changes that needed to happen as I went through the process. Now that we have the data we shouldn't ever need to look at this code again.

fireproj_trainml_and_predict will train the model on all the data we have available, pull the most recent hourly data and convert everything to the same format as the daily data in our archive, make a prediction based on the most recent hourly data, and then output a map of our best prediction for today as well as any fire spots. The fire spots are red 1 km squares. If there is no fires there will be a marking saying so. The prediction will be blue for we're not predicting a fire, and orange for we are predicting a fire. The size of the prediction circle is proportionate to our confidence level. 

fireproj_updatedatadaily can update our stored data so it can be ran once a day and it'll pick up any previous dates of fire and weather data. 

This should be everything we need. Feel free to tinker with it, especially the ML model since it is not great. The model is overfitted to 0 since Maui did not have any fire days.